+++ docs/changelog.rst
@@ -1,7 +1,7 @@
 0.5.5
 -----
 
-- added option to define custom templates through option ``--templates_path``, read more in :ref:`tutorial section <custom-templates>`
+- added option to define custom templates through option ``--templates_path`` and ``--templates_theme``, read more in :ref:`tutorial section <custom-templates>`
 - url parameter can also be an Engine instance (this usage is discouraged though sometimes necessary)
 - added support for SQLAlchemy 0.6 (missing oracle and firebird) by Michael Bayer
 - alter, create, drop column / rename table / rename index constructs now accept `alter_metadata` parameter. If True, it will modify Column/Table objects according to changes. Otherwise, everything will be untouched.