+++ docs/changelog.rst
@@ -3,7 +3,7 @@
 - added option to define custom templates through option ``--templates_path`` and ``--templates_theme``, read more in :ref:`tutorial section <custom-templates>`
 - use Python logging for output, can be shut down by passing ``--disable_logging`` to :func:`migrate.versioning.shell.main`
 - `url` parameter can also be an :class:`Engine` instance (this usage is discouraged though sometimes necessary)
-- added support for SQLAlchemy 0.6 (missing oracle and firebird) by Michael Bayer
+- added support for SQLAlchemy 0.6 by Michael Bayer
 - alter, create, drop column / rename table / rename index constructs now accept `alter_metadata` parameter. If True, it will modify Column/Table objects according to changes. Otherwise, everything will be untouched.
 - complete refactoring of :class:`~migrate.changeset.schema.ColumnDelta` (fixes issue 23)
 - added support for :ref:`firebird <firebird-d>`